+++ test/option-short.js
@@ -6,13 +6,16 @@ describe('short options', function() {
         const command = cli.command('test')
             .option('-f, --foo', 'Foo')
             .option('-b, --bar', 'Bar')
-            .option('-x, --baz', 'Baz');
+            .option('-x, --baz', 'Baz')
+            .option('-0, --zero', 'Zero');
 
         [
-            { test: '-f', expected: { foo: true, bar: false, baz: false } },
-            { test: '-fb', expected: { foo: true, bar: true, baz: false } },
-            { test: '-fbx', expected: { foo: true, bar: true, baz: true } },
-            { test: '-xfbfx', expected: { foo: true, bar: true, baz: true } }
+            { test: '-f', expected: { foo: true, bar: false, baz: false, zero: false } },
+            { test: '-0', expected: { foo: false, bar: false, baz: false, zero: true } },
+            { test: '-fb', expected: { foo: true, bar: true, baz: false, zero: false } },
+            { test: '-f0b', expected: { foo: true, bar: true, baz: false, zero: true } },
+            { test: '-fbx', expected: { foo: true, bar: true, baz: true, zero: false } },
+            { test: '-xfbfx', expected: { foo: true, bar: true, baz: true, zero: false } }
         ].forEach(testcase =>
             it(testcase.test, () => {
                 const actual = command.run([testcase.test]);
@@ -30,7 +33,7 @@ describe('short options', function() {
             it(test, () =>
                 assert.throws(
                     () => command.run(['-fz']),
-                    /Unknown short option: -z/
+                    /Unknown option "z" in short option sequence: -fz/
                 )
             );
         });
@@ -43,11 +46,11 @@ describe('short options', function() {
 
         assert.throws(
             () => command.run(['-fb']),
-            /Non-boolean option -b can't be used in short option sequence: -fb/
+            /Non-boolean option "-b" can't be used in short option sequence: -fb/
         );
         assert.throws(
             () => command.run(['-bf']),
-            /Non-boolean option -b can't be used in short option sequence: -bf/
+            /Non-boolean option "-b" can't be used in short option sequence: -bf/
         );
     });
 });